Adding method to remove the directory with its content.

+++ admin/class-import.php
@@ -181,11 +181,37 @@ class WPSEO_Import {
 		if ( file_exists( $this->filename ) && is_writable( $this->filename ) ) {
 			unlink( $this->filename );
 		}
-		if ( file_exists( $this->file['file'] ) && is_writable( $this->file['file'] ) ) {
+		if ( !empty( $this->file['file'] ) && file_exists( $this->file['file'] ) && is_writable( $this->file['file'] ) ) {
 			unlink( $this->file['file'] );
 		}
 		if ( file_exists( $this->path ) && is_writable( $this->path ) ) {
-			rmdir( $this->path );
+			$this->remove_directory( $this->path );
 		}
+	}
+
+	/**
+	 * Removes a directory with its content.
+	 *
+	 * @param string $directory The directory to remove.
+	 */
+	private function remove_directory( $directory ) {
+		$current_directory = opendir( $directory );
+		while ( $filename = readdir( $current_directory ) ) {
+
+			if ( $filename === '.' || $filename === '..' ) {
+				continue;
+			}
+
+			if ( is_dir( $directory . '/' . $filename ) ) {
+				$this->remove_directory( $directory . '/' . $filename );
+			}
+
+			if ( is_file( $directory . '/' . $filename ) ) {
+				unlink( $directory . '/' . $filename );
+			}
+		}
+
+		closedir( $current_directory );
+		rmdir( $directory );
 	}
 }
